refactor(gsw): read truecolor channels with one shared scanner
